change ratio to int64 to prevent unsafe type in crd

+++ internal/quota/quotalists.go
@@ -95,7 +95,7 @@ func (pcr QuotaLists) Min() Quotas {
 	return quotaResources
 }
 
-func (pcr QuotaLists) OptimalValues(ratio float64, minQuotas Quotas, maxQuotas Quotas) Quotas {
+func (pcr QuotaLists) OptimalValues(ratio int64, minQuotas Quotas, maxQuotas Quotas) Quotas {
 	// Calculate resources with 3 different approaches and select largest value
 	approaches := NewQuotaLists()
 	approaches.Append(pcr.Sum().Resized(ratio))
